trip_planner_agent/requirements.txt
crewai
streamlit 
openai
unstructured
pyowm
tools
langchain_google_genai
python-dotenv==1.0.1
duckduckgo-search
crewai[tools]
langchain